+++ tests/nexus.lua
@@ -4,6 +4,7 @@ function start()
   Add(15, 7.5, Peg.Level, 0, "missions/sewer1", "Flood")
   Add(5, 11, Peg.Level, 0, "missions/state-lab", "State Lab")
   Add(10, 11, Peg.Level, 0, "missions/render-lab", "Render Lab")
+  Add(15, 11, Peg.Level, 0, "missions/paradox-lab", "Paradox Lab")
   Add(4, 4, Peg.Back)
   return { bg = {0.1, 0.04, 0.45} }
 end